Refactor floating gum to collectibles
- To add new collectible, add new name for object in Tiled, add new Json info, and add new collision behavior
