+++ core/src/edu/cornell/gdiac/bubblegumbandit/models/level/LevelModel.java
@@ -15,11 +15,9 @@
 
 package edu.cornell.gdiac.bubblegumbandit.models.level;
 
-import com.badlogic.gdx.graphics.Camera;
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
-import com.badlogic.gdx.math.Circle;
 import com.badlogic.gdx.physics.box2d.*;
 
 import com.badlogic.gdx.math.Rectangle;
@@ -31,7 +29,6 @@ import edu.cornell.gdiac.bubblegumbandit.controllers.AIController;
 import edu.cornell.gdiac.bubblegumbandit.helpers.TiledParser;
 import edu.cornell.gdiac.bubblegumbandit.models.enemy.EnemyModel;
 import edu.cornell.gdiac.bubblegumbandit.models.enemy.MovingEnemyModel;
-import edu.cornell.gdiac.bubblegumbandit.models.level.gum.FloatingGum;
 import edu.cornell.gdiac.bubblegumbandit.models.player.BanditModel;
 import edu.cornell.gdiac.physics.obstacle.Obstacle;
 import edu.cornell.gdiac.util.PooledList;
@@ -316,7 +313,7 @@ public class LevelModel {
                     }
                     break;
                 case "Gum":
-                    FloatingGum gum = new FloatingGum();
+                    Collectible gum = new Collectible();
                     gum.initialize(directory, x, y, scale, constants.get("floatingGums"));
                     activate(gum);
                     break;
